fix badge coverage; add example instance
