+++ README.md
@@ -1,6 +1,6 @@
 # norris-nrti
 
 [![Build Status](https://travis-ci.org/DeltaGraphs/norris-nrti.svg?branch=development)](https://travis-ci.org/DeltaGraphs/norris-nrti)
-[![Coverage Status](https://coveralls.io/repos/DeltaGraphs/norris-nrti/badge.svg)](https://coveralls.io/r/DeltaGraphs/norris-nrti)
+[![Coverage Status](https://coveralls.io/repos/DeltaGraphs/norris-nrti/badge.svg?branch=development)](https://coveralls.io/r/DeltaGraphs/norris-nrti?branch=development)
 
 Node Real-Time Intelligence (Credits CoffeeStrap)
