--- test/.jshintrc
@@ -1,39 +1,0 @@
-{
-  "curly": true,
-  "eqeqeq": true,
-  "eqnull": true,
-  "expr": true,
-  "latedef": false,
-  "onevar": false,
-  "noarg": true,
-  "trailing": true,
-  "undef": true,
-  "asi": true,  
-  "unused": true,
-  "evil": true,
-  "strict": true,
-  "sub": false,
-  "regexdash": true,
-  "wsh": true,
-  "immed": true,
-  "white": false,
-  "bitwise": true,
-  "camelcase": true,
-  "indent": 4,
-  "newcap": false,
-  "noempty": true,
-  "nonew": true,
-  "plusplus": false,
-  "quotmark": "single",
-  "asi": false,
-  "smarttabs": false,
-  "boss": true,
-  "predef": [
-    "describe",
-    "it",
-    "before",
-    "beforeEach",
-    "after",
-    "afterEach"
-  ]
-}
